index.html: link to alyssa.is and TWIS
It was recently pointed out to me[1] that it's not really clear from
the website that Spectrum is being actively developed, or what point
it is at in its development.

To try to make this clearer, I've added information about blog posts
and email updates to the homepage, so people coming across it should
be able to easily see that the weekly updates are indeed being posted,
and progress is being made.

[1]: https://mastodon.social/@rysiek/104105383686173382