bibliography.html: create
I thought it might be useful to have a place to collect links that are
particularly relevant to Spectrum.  This serves two purposes: to
provide further external reading for people learning about Spectrum
from the website; and to make it easy for us to find links we refer to
often.

At some point I might add author names, etc. to these links, and make
them a bit more like proper citations, but for now I wanted to get an
initial version of this out there rather than bikeshedding citation
formats.

Message-Id: <20200726045701.32259-1-hi@alyssa.is>
Reviewed-by: Cole Helbling <cole.e.helbling@outlook.com>