Fix index links on top-level pages
Linking to .. means linking to the directory above these pages, when
we actually want to link to the directory these pages are in (which
should display index.html).  This doesn't have any effect on
<https://spectrum-os.org/>, because the site is already at the top
level, but it would cause a problem if the site was elsewhere -- for
example, when viewing it locally while working on it.

Message-Id: <20200726055410.20641-1-hi@alyssa.is>
Reviewed-by: Cole Helbling <cole.e.helbling@outlook.com>