design.html: backpedal a bit on POWER9
Supporting POWER9 isn't going to be as easy as I'd initially hoped.
This patch aims to explain that there are blockers that I'd need help
overcoming to be able to support it, even though I'd still like to.

I don't want to be promising anything I can't deliver on, and I now
know that it's very unlikely I'll be able to deliver on this without
help.

Reported-by: Molly Miller <mm@m-squa.red>
Message-Id: <20201118120116.21390-1-hi@alyssa.is>