index.html: add alt text to logo
The empty alt text would make the image basically invisible if the
user agent didn't display the image for whatever reason.  This would
have been fine if the image was entirely decorative, but since it's
also the only way to navigate to the logo page it's important it has
some fallback content so there's still something to select.

Message-Id: <20201219142546.21683-1-hi@alyssa.is>
Reviewed-by: Cole Helbling <cole.e.helbling@outlook.com>