Update Postorius URLs
These changed with a mailman-web update.  I've added redirects from
the old locations but it's still better to avoid them.

The update also wanted to change /lists/hyperkitty to /lists/archives,
but on spectrum-os.org that path is used for public-inbox, so I've
patched mailman-web to keep that at /lists/hyperkitty.

Message-Id: <20210414161133.14590-1-hi@alyssa.is>
Reviewed-by: Cole Helbling <cole.e.helbling@outlook.com>