style.css: highlight targets
On pages like participating.html, it was unclear what was happening
when somebody clicked on a link that went to a section near the bottom
of the page — the page would jump down, but it wasn't obvious what
part in particular they were supposed to be looking at.

So let's highlight them to make it really obvious.

Signed-off-by: Alyssa Ross <hi@alyssa.is>