Update descriptions of the project and system
There hasn't been a This Week in Spectrum email in ages, but there are
now the regular development streams, so it makes more sense to point
there as a way to follow current developments in the project.

The descriptions of the system had become very out of date, having
mostly been written during preliminary stages before I even began
putting the Spectrum system together, and at the same time, were very
specific in how they envisaged the system working in ways I'm no
longer sure about.  So I've updated those descriptions to reflect
current reality, and reduced the focus on things that are not
implemented (or at least designed) yet.

Signed-off-by: Alyssa Ross <hi@alyssa.is>