host/rootfs: remove ext-vmms dependency from vm-env
The only difference between services being listed in vm-env or ok-all
is whether vm-start will wait for it, but there's no need for vm-start
to wait for every VMM to be up — just the one it's interested in (and
of course any of its providers).  As we move towards multiple sources
of VMs, it makes increasingly little sense to even have a concept of
all VMMs being up.

Signed-off-by: Alyssa Ross <hi@alyssa.is>