host/initramfs/extfs.nix: remove
We shouldn't be distributing a pre-formatted mutable partition in
images, because it can't be encrypted.  There's also no longer any
special code in Spectrum for handling mutable partitions, so we don't
need it in most of the test setups, as evidenced by the fact that it
was never writable in most of them anyway.  Where that is
useful (host/rootfs), we can just format a new partition as part of
the creation of a mutable image we already have.

Signed-off-by: Alyssa Ross <hi@alyssa.is>