tools: combine Meson projects
This allows having a single set of wraps shared by all first-party
programs in Spectrum, and it will also mean that, if people want to
use custom VMs, they just need to install a single package to get all
Spectrum integrations.

For now, since there's exactly one host package and one guest package,
I've left the Nix packaging mostly as-is, but the next step will be to
change it to just have guest and host packages rather than a package
for each program.

Signed-off-by: Alyssa Ross <hi@alyssa.is>