tools/start-vmm: name network devices with VM ID
This is a step towards not having to create and delete the devices
every VM boot, by giving them a name that continues to make sense
between VMM invocations.

Cloud Hypervisor IDs for devices attached to the provider VM are still
process-group-based, because otherwise we'd be very likely to run into
an annoying race condition: Cloud Hypervisor's vm.remove-device API
endpoint returns before the device has been ejected, so the ID isn't
available for immediate reuse.  Because these IDs no longer correspond
to the name of a network device, I've changed the prefix from "router"
to "client", because from the perspective of the router, these are
devices for talking to clients.  I've also changed struct net_config's
id member to be the full possible length of "client", followed by a
process group ID, followed by a NUL terminator, now that we no longer
have to inherit the maximum size from interface names.

Signed-off-by: Alyssa Ross <hi@alyssa.is>