tools/start-vmm: notify readiness earlier
Readiness, in the context of the vmm service, means that it's the API
socket is ready for communication from other services.  The VM has to
be created before anything else useful can be done with the API, but
it's not necessary for all the network devices to have been set up.
In fact, separating those two steps further will be necessary to fix
a current race condition where there's no guarantee a network provider
VMM will be ready by the time a client tries to add a network device
to it.

Signed-off-by: Alyssa Ross <hi@alyssa.is>