host/rootfs: let the kernel number dead devices
These could stick around for a while.  Doing it this way means we
don't have to worry about conflicts if pids are reused.

Busybox ip gets confused by the name of the interface changing to
something other than what we gave it, so switch to iproute2, which
handles this correctly.

Signed-off-by: Alyssa Ross <hi@alyssa.is>