tools/start-vmm: use persistent TAPs
Now that these have predictable names, we can reuse them between VM
boots, which will reduce the complexity of starting a VM.

As a result, we can no longer use the process group of the VMM process
to generate the MAC address, because if a client VM is restarted, the
router VM will be expecting to send to an IP address based on the
process group of the previous VMM instance, but the client will have
configured itself based on the new process group.  Instead, we now use
the index of the client's network interface, since that will now never
change.  An interface index is four bytes, so our MAC addresses are
adjusted to be able to fit the whole thing, and also to conform to IEE
802c while we're here.  We can't fit all four bytes in the IPv4
addresses derived from the MAC addresses, but that's only going to be
a problem after the 21845th VM, so I'm not too worried for now.

Signed-off-by: Alyssa Ross <hi@alyssa.is>