vm: use passt for development VM networking
This allows "make run" development VMs to have working, unprivileged
networking.  It isn't being used with crosvm yet because crosvm
requires some patches[1][2], and it isn't being used with QEMU in
vm/sys/net, because that network device is e1000e for realism, and
QEMU only supports vhost-user with virtio-net.

Link: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6491650 [1]
Link: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6491651 [2]
Signed-off-by: Alyssa Ross <hi@alyssa.is>