release/live: drop UKI addon hack
There's no longer a hard limit of 96 PE sections on UKIs, so we can
build all the device trees we have into the main image, without
needing loads of addons.

This increases the amount of RAM required to get past the stub, hence
the increase in the try test, but 1G is still a very reasonable amount
of RAM to require for a virtualization-based system.

Link: https://github.com/systemd/systemd/commit/8c5b359579b0f1029edafb0bd96b5ebfb271db76
Signed-off-by: Alyssa Ross <hi@alyssa.is>