host/rootfs: poll active consoles for serial getty
Previously, the active consoles were only checked once, so if a
console appeared later, getty wouldn't be run for it.  This meant that
if the console was set to a serial device whose driver was not
built-in, it might have no shell if the module was not loaded until
after rc.init had already run.  Also, if a console device was removed,
its instance of the serial-getty would hang around.

To fix this, replace the getty instantiation in rc.init with a service
that will create and destroy serial-getty instances to match the
currently configured consoles, which will rerun whenever the
configured consoles change.

Fixes: b096279 ("host/rootfs: run getty for active serial consoles")
Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-ID: <20250613140508.76998-2-hi@alyssa.is>