release/checks/integration: use default Meson timeout
The failure mode for integration tests is usually going to be that
they get stuck, so it's important for the build to fail in that case.

The timeout of 240 seconds was determined by rounding up and doubling
the longest time I observed on my x86_64 laptop under load (107
seconds).  The multiplier value of 15 was determined by
unscientifically by dividing how long it normally takes my aarch64
laptop to run the test by how long it takes my x86_64 one.

Signed-off-by: Alyssa Ross <hi@alyssa.is>