release/checks/integration: test late serial
By "late", I mean the serial device won't be configured by the kernel
until full userspace is running.  Here I'm doing this with a USB
serial device, for which driver modules need to be loaded by the
kernel.  This is a regression test, because that didn't used to work,
since serial consoles were only configured once at boot.

Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-ID: <20250613140508.76998-9-hi@alyssa.is>