img/app: Run PipeWire and WirePlumber in the VMs
WirePlumber is completely overkill as a session manager here, and
ideally a trivial session manager would be used instead.  PipeWire is
configured to listen on the PulseAudio socket, so PulseAudio
compatibility works.  pw-record and pw-play both work, and if PulseAudio
is installed paplay and parecord also work.  This does install a lot of
unnecessary files into the VMs, which will hopefully be removed
later as part of a debloating effort.

Only run-qemu has had a virtio-sound device added, as crosvm and Cloud
Hypervisor require a virtio-user sound device and that is more complex
to set up.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>
---
Changes since v8:
- Add longer explanation for the priority.driver and priority.session
  values.
- Give credit to George Kiagiadakis.
- Fix spelling errors in comments.

Changes since v7:
- Give the capture node a higher priority.session than the playback
  node, so WirePlumber links recording applications to the capture node
  instead of the monitor of the playback node.
- Give the capture node a higher priority.driver than the playback node,
  so PipeWire has the capture node drive the graph.  This is better
  because capture xruns lead to permanently corrupted data, whereas
  playback xruns do not.
- Re-enable monitor node creation in WirePlumber.

Changes since v6:
- Fix spelling errors in commit messages.
- Sort lines in Makefile.
- Don't disable support.settings as a comment in wireplumber.conf says
  to not do that.  Instead, tell WirePlumber to not create monitor
  nodes.  This is a workaround for WirePlumber bug 829.
- Don't remove "-cpu host" from make run-qemu's QEMU command line.  This
  was needed for local testing with KVM disabled but should not have
  been submitted upstream.  Hardware with KVM support should be used
  instead.

Changes since v5:

- Use a drop-in configuration file for WirePlumber instead of overriding
  wireplumber.conf.  This should be more robust to future WirePlumber
  changes.