img/app: Run PipeWire and WirePlumber in the VMs
WirePlumber is completely overkill as a session manager here, and
ideally a trivial session manager would be used instead.  PipeWire is
configured to listen on the PulseAudio socket, so PulseAudio
compatibility works.  pw-record and pw-play both work, and if PulseAudio
is installed paplay and parecord also work.  This does install a lot of
unnecessary files into the VMs, which will hopefully be removed
later as part of a debloating effort.

If WirePlumber is allowed to enable monitor nodes for the ALSA output
device, WirePlumber prefers the monitor over the ALSA input device [1].
This breaks audio recording.  To work around this, set
node.features.audio.monitor-ports to false.

Only run-qemu has had a virtio-sound device added, as crosvm and Cloud
Hypervisor require a virtio-user sound device and that is more complex
to set up.

[1]: https://gitlab.freedesktop.org/pipewire/wireplumber/-/issues/829

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>
---
Changes since v6:

- Fix spelling errors in commit message.
- Sort lines in Makefile.
- Don't disable support.settings as a comment in wireplumber.conf says
  to not do that.  Instead, tell WirePlumber to not create monitor
  nodes.  This is a workaround for WirePlumber bug 829.
- Don't remove "-cpu host" from make run-qemu's QEMU command line.  This
  was needed for local testing with KVM disabled but should not have
  been submitted upstream.  Hardware with KVM support should be used
  instead.

Changes since v5:

- Remove "directories" service in favor of creating the directories from
  rc.init.
- Use a drop-in configuration file for WirePlumber instead of overriding
  wireplumber.conf.  This should be more robust to future WirePlumber
  changes.