img/app: Run PipeWire and WirePlumber in the VMs
WirePlumber is completely overkill as a session manager here, and
ideally a trivial session manager would be used instead.  PipeWire is
configured to listen on the PulseAudio socket, so PulseAudio
compatibility works.  pw-record and pw-play both work, and if PulseAudio
is installe paplay and parecord also work.  This does install a lot of
unnecessary files into the VMs, which will hopefully be removed
later as part of a debloating effort.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>