tools/start-vmm: let s6 handle network cleanup
This simplifies start-vmm a lot.  This does mean that cleanup wouldn't
happen for AppImage VMs, since s6 doesn't run those, but we don't
currently support networking for those anyway, and I'm hoping to
further clean up networking by then such that manual cleanup is
unnecessary.

The 4th argument that s6 gives to the finish script is the process
group ID, not the pid of the main process.  These should always be the
same, but for correctness start-vmm is modified to use the process
group ID for naming devices.

Signed-off-by: Alyssa Ross <hi@alyssa.is>