host/start-vmm: fix double_fork() error returns
The comment says it's supposed to return a negative errno value on
failure, but it was only doing this in the case where everything went
right in the parent process.  If pipe or fork failed, it returned -1,
and if something went wrong when reading, it returned the buffer it
was reading into, which at that point I think was likely still
uninitialized, at least partially.

Fixes: a2d362c ("host: allow VMs to be powered off")
Signed-off-by: Alyssa Ross <hi@alyssa.is>