release/checks/integration: always be draining
If QEMU ever can't write to the console, it will block.  So if we're
ever writing to the console, while QEMU is trying to write, there's a
chance of deadlock.  The easiest way to avoid that is to make sure we
are always draining the console on its own thread.

The only reading from the console the tests currently do is waiting
for a prompt, so the console-draining thread is always scanning for a
prompt and will write into a pipe if one is found.  The pipe is
non-blocking to ensure the thread never blocks.

Fixes: ae45978 ("release/checks/networking: init")
Signed-off-by: Alyssa Ross <hi@alyssa.is>