release/checks/integration: unlock console reads
When the process exits (in this case because the main thread has
reached the end of the test), libc will clean up all open streams,
which requires locking them.  If the reader thread is in fgetc at that
point, there will be a deadlock because the reader thread will hold
the lock until it receives a character, but it's likely that no more
characters will ever be received at that point.

To avoid this, don't lock in the reader thread.  Since the stream is
only ever read from on the reader thread, this should be fine.

Fixes: 048d403 ("release/checks/integration: always be draining")
Signed-off-by: Alyssa Ross <hi@alyssa.is>