scripts/run-spectrum-vm.c: remove
This basically required reimplementing the Spectrum host root
filesystem, which was a lot to maintain, and running it requires KVM,
which we want to avoid in Nix builds.  Now that we have integration
tests (which can run without KVM), this isn't really testing anything
not covered by them anyway, so let's just remove it.

Signed-off-by: Alyssa Ross <hi@alyssa.is>