Revert "img/app: fix permissions on /tmp"
The default permissions of 1777 are correct, as `stat /tmp` on any
Unix-like system will show.  Using 1755 is nonsensical: it means that
items in the directory can only be moved or deleted by their owners, but
also that they can only be moved or deleted by the owner of the
directory.  In this case, that owner is root, who can ignore permissions
anyway.

This reverts commit 306350cf1cd85c3f3c890c076e989a2cd5ca6d9e.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>
Message-ID: <6d322122-a8da-4762-ba25-6b981eab01eb@gmail.com>
Reviewed-by: Alyssa Ross <hi@alyssa.is>
Tested-by: Alyssa Ross <hi@alyssa.is>
Signed-off-by: Alyssa Ross <hi@alyssa.is>