img/app: Create needed directories in early boot
This moves various calls to mkdir(1) to very early boot, before any
services are running.  This has two advantages:

1. These directories are guaranteed to exist.  Code can just assume that
   they are there without checking for them.

2. Malicious code running as an unprivileged user cannot create
   directories under /tmp before legitimate code has done so.

Also, it creates the various directories used by X11 with restrictive
permissions to prevent untrusted code from writing to them, and sets up
/run/user/0 to provide $XDG_RUNTIME_DIR.

The copyright notice for direcatory creation is not kept because making
four directories with well-known names and permisisons is not
copyrightable.