img/app: run PipeWire and WirePlumber in the VMs
WirePlumber is completely overkill as a session manager here, and
ideally a trivial session manager would be used instead.  PipeWire is
configured to listen on the PulseAudio socket, so PulseAudio
compatibility works.  pw-record and pw-play both work, and if PulseAudio
is installed paplay and parecord also work.  This does install a lot of
unnecessary files into the VMs, which will hopefully be removed
later as part of a debloating effort.

Only run-qemu has had a virtio-sound device added, as crosvm and Cloud
Hypervisor require a vhost-user sound device and that is more complex
to set up.

PipeWire and WirePlumber require that "$XDG_RUNTIME_DIR" exist, so
create it in rc.init.  Having a separate service to create a single
directory is not worthwhile.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>
Message-ID: <20250731-pipewire-v10-v11-1-90ad150e2c00@gmail.com>