img/app: Create X11 directories in early boot
They are created with mode 1700, which ensures that no untrusted code
can ever create them later and spoof the actual X11 server.  Some code
checks for the owner being root and the sticky bit being set, so 0700
might break some things.  This does break Inter-Client Exchange (ICE),
X Input Method (XIM), and X Font Server (XFT).  However, these are only
used by software that Spectrum VMs don't run.  Abstract sockets are a
hole in this, but I have sent a patch to the Linux kernel to disable
connecting to them and a patch to libxcb to not use them.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>