Add /dev/fd and /dev/std*
This is nearly same as 14483e1a690c (img/app: add /dev/fd and /dev/std*),
but for the host and for vm/sys/net.  The only difference is that the
symlinks are created in init, rather than in rc.init.  This ensures that
nothing in the main system can run before they the links are created.

While only Spectrum-provided code should run in these VMs, third-party
dependencies are likely to only be tested in an environment where these
links are present.  Since the BSDs also have these links, some code
might even use them on purpose.

Not having these links could cause severe bugs.  For instance, if
/dev/stdout or /dev/stderr is missing, writing to them will create a
regular file instead.  This could cause other code to misbehave.
Furthermore, /dev is a devtmpfs, so this could consume a lot of memory.

[1]: https://spectrum-os.org/lists/archives/spectrum-devel/87h5xdjo9p.fsf@alyssa.is/

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>