Add /dev/fd and /dev/std*
This is the same as 14483e1a690c (img/app: add /dev/fd and /dev/std*),
but for the host and for vm/sys/net.  While only Spectrum-provided code
should run in these VMs, third-party dependencies of Spectrum might
assume these links exist, and them being missing could cause severe
bugs.  For instance, code writing to /dev/stdout could create a file in
/dev rather than actually writing to stdout.

In the host, the links are added in the initramfs.  Since /dev is
created by the kernel and moved (via mount --move) from the initramfs to
the main system, adding the links in the main system is not necessary
and in fact would fail.

Also reorder the moving of /sys, /proc, and /dev from the initramfs to
the root filesystem to minimize the time that /dev and /proc are not
mounted.  /proc is considered more important than /dev.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>