Many image fixes and systemd integration
Patches 1 through 19 are all fixes or enhancements to the image build
process.  There are other changes that need to be done around error
handling, but these are all useful regardless.  See the individual
commit messages for details.

Notably, one of these patches standardizes file modes so that they are
not dependent on the permissions in the user's git repository (except
for whether the executable bit is set, which git stores).  This is
because that depends on things like the user's umask, and thus should
have no effect on the image.

Patch 20 switches from s6-linux-init to systemd.  This is not intended
for merging, at least not yet.  However, it *is* meant to show the
beginning of how Spectrum could benefit from systemd's features.
Notably, this patch reduces the amount of code.  This is despite all
Spectrum-specific services still being managed by s6 and additional
complexity in the Nix files being needed to work around nixpkgs not
using standard directories to find things like systemd unit files and
PAM modules.  It's also worth noting that at least GNOME has a fairly
hard dependency on systemd, but I doubt COSMIC will as parts of it are
even used on Redox, which definitely does not run systemd!

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>

---
Changes in v2:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v1: https://lore.kernel.org/r/20250904-systemd-v1-0-2a63b790a913@gmail.com

--- b4-submit-tracking ---
{
  "series": {
    "revision": 2,
    "change-id": "20250815-systemd-2cdd0b578a86",
    "prefixes": [],
    "history": {
      "v1": [
        "20250904-systemd-v1-0-2a63b790a913@gmail.com"
      ],
    }
  }
}