host/rootfs: Switch to systemd
This requires removing the s6 calls to getty (now handled by systemd)
and the use of mdevd (replaced by systemd-udevd).  Additionally,
s6-svscan is called by systemd instead of by s6-linux-init, and
/run/service is populated by systemd-tmpfiles instead of by
s6-linux-init.

This overall reduces the amount of code, as systemd does so much itself
and thus Spectrum OS does not need to reimplement as much.  Furthermore,
more savings and additional features could be obtained by using
more of systemd.  For instance, weston could be launched by a systemd
service instead of s6, meaning that s6 would only be used to launch
the per-VM services.  Furthermore, the lifetime of the login session
could be tied to the lifetime of the current process, so that when the
user logs out (or their session is otherwise terminated, perhaps by
Linux's SAK killing the compositor's parent process) all of their VMs
are killed.  Finally, some sandboxing features are trivial to implement
with systemd.  For instance, host processes are forbidden from using
Linux kernel IP networking: they can configure interfaces as normal, so
guest networking works, but they cannot send or receive any packets.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>