Standardize directories and symlinks in images
There are a few directories and symbolic links that a Linux system
should always have.  Even if Spectrum OS itself does not use them,
third-party dependencies and/or applications might rely on them.
Create these in scripts/make-erofs.sh rather than separately in
each VM's build scripts.  The creation of /run/lock assumes that
s6-linux-init is being used, but that assumption is easy to fix later.
This also enforces that the symlinks and directories were *not* created
in other places.  The app VM build violated this rule, so fix it.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>