Switch from mdevd to systemd-udevd in root filesystem
systemd-udevd provides support for significantly more hardware than
mdevd does.  Therefore, switch to it.

The last patch intentionally uses some packages built with glibc to
avoid having to build the world from source.  The use of glibc isn't
essential and is just because not using the binary cache to build is
very, very slow.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>

---
Changes in v2:
- Use more optional arguments instead of positional parameters in
  sd-notify-wrapper.
- Explain why explicit modprobe is needed.
- Link to v1: https://spectrum-os.org/lists/archives/spectrum-devel/20250913-udev-v1-0-eade4ab8f2b4@gmail.com

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20250911-udev-24cf4bf68fbe",
    "prefixes": [],
    "history": {
      "v1": [
        "20250913-udev-v1-0-eade4ab8f2b4@gmail.com"
      ]
    }
  }
}