tools: Add adapter tool for services using sd_notify
This adapts programs using sd_notify for use with s6 readiness
notification.

I chose to use Linux-specific epoll(7) and signalfd(2).  It avoids
having to deal with asynchronous signal handlers and overall makes the
code simpler and more readable.  Nevertheless, the code is still quite
complicated.  This is partly because of how ugly the POSIX interface is,
but also because the code uses no third-party dependencies.  Using
skalibs might make the code simpler at the cost of making it harder for
those not familiar with skalibs to understand.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>
---
The amount of code in sd-notify-adapter is less than that of
s6-supervise itself.  Furthermore, it duplicates a lot of the
functionality in that tool.  Therefore, I think that s6-supervise is
actually the correct place for this.  In particular, this allows
reliable reloading of services: send a SIGHUP and wait for RELOADING=1
and then READY=1.