tools: Add adapter tool for services using sd_notify
This adapts programs using sd_notify for use with s6 readiness
notification.

I chose to use Linux-specific epoll(7) and signalfd(2).  It avoids
having to deal with asynchronous signal handlers and overall makes the
code simpler and more readable.  Nevertheless, the code is still quite
complicated.  This is partly because of how ugly the POSIX interface is,
but also because the code uses no third-party dependencies.  Using
skalibs might make the code simpler at the cost of making it harder for
those not familiar with skalibs to understand.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>
---
systemd readiness notification has two
strict advantages over the s6 version:

1. It allows reliable reloading.
2. It allows providing a status message that the service manager
   can show in status output.

s6 would actually benefit from both of these features.

This code also does a lot of validation on provided file descriptors.
It would be simpler to just check that the first file descriptor opened
is at least 3 and that ioctl(notification_fd, FIOCLEX) succeeds.  This
checks for all file descriptor numbers being used, but *not* for stdin
being unreadable or stdout/stderr/notification_fd being unwriteable.