tools/sd-notify-adapter: Close leaked FDs in parent after fork()
It necessary to close all file descriptors other than stdin, stdout, and
stderr that were inherited from the parent, and to point stdin and
stdout at /dev/null.  Otherwise the child will keep file descriptors
open that it should not, and that's an observable bug.

An s6-supervise lock FD must not be closed, though.  Add a --lock-fd
option to exempt the provided FD from the closing spree.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>