tools/sd-notify-adapter: Support adjusting OOM weight
systemd-udevd should never be killed by the OOM killer, so this is
useful.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>
---
The remainder of the patches in this series do not use this feature.  It
would be better to set the OOM weight before spawning init, and
explicitly enable the OOM killer in child processes via the run script.
It would be even better for s6-supervise to set its own OOM weight to
-1000 on its own.