host/rootfs: Switch to systemd-udevd
This replaces the mdevd uevent daemon with systemd-udevd, which supports
much more hardware and is therefore to be preferred for a full desktop
environment.  Various Wayland compositors integrate with systemd-udevd,
allowing them to discover devices as they appear rather than having to
only use devices plugged in when the compositor is started.
Additionally, systemd-udevd has quirks for various devices that are
needed to support the wide variety of hardware end-users have.

The dependency on /dev/dri/card0 being present is eliminated, and
whatever devices the user has are now picked up by the compositor.  New
dependencies are added to ensure that udev coldplug has finished before
any non-trivial services are started.  systemd-udev-trigger.service runs
'udevadm trigger' and has Before=sysinit.target, so anything that is not
an early boot service can assume 'udevadm trigger' has run.

systemd-udevd unsets PATH.  Therefore, manually set it to /usr/bin
before invoking any scripts, like /usr/libexec/net-add.

systemd-udevd provides the 'udevadm wait' command to wait for a device
to appear and be processed.  This replaces all uses of /etc/mdev/wait,
so this script is removed.  /etc/mdev/listen and /run/wait only existed
to wake up /etc/mdev/wait, so they are also removed.

systemd-udevd does have a module loading feature via the 'kmod load'
builtin.  However, this does not work, so a manual call to modprobe is
used instead.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>
---
I am not sure why the 'kmod load' builtin is broken in Spectrum.  I
suspect it is either a bug in Nixpkgs or a bug in systemd-udevd when not
spawned by systemd.  The simplest way to check which is the case would
be to see what Gentoo does, as Gentoo uses systemd-udevd with OpenRC by
default.