host/rootfs: Switch to systemd-udevd
This replaces the mdevd uevent daemon with systemd-udevd, which supports
much more hardware and is therefore to be preferred for a full desktop
environment.  Various Wayland compositors integrate with systemd-udevd,
allowing them to discover devices as they appear rather than having to
only use devices plugged in when the compositor is started.
Additionally, systemd-udevd has quirks for various devices that are
needed to support the wide variety of hardware end-users have.

The dependency on /dev/dri/card0 being present is eliminated, and
whatever devices the user has are now picked up by the compositor.  New
dependencies are added to ensure that udev coldplug has finished before
any non-trivial services are started.  systemd-udev-trigger.service runs
'udevadm trigger' and has Before=sysinit.target, so anything that is not
an early boot service can assume 'udevadm trigger' has run.

systemd-udevd doesn't set PATH to anything useful, presumably because
under NixOS this is handled some other way.  Therefore, explicitly set
it to /usr/bin in the scripts systemd-udevd calls.

The /etc/mdev/listen and /etc/mdev/wait scripts are replaced by
'udevadm wait', so they are deleted.  The whole static-nodes target also
goes away, as Linux handles that itself and systemd-udevd deals with
autoloading.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>
---
The changes to default.nix are so that I could use the binary cache,
which was necessary to test this in a reasonable amount of time.  I'm
also not sure what changes you (Alyssa) want in this file.  Feel free to
fix it up on commit.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>