host/rootfs: Avoid redundant mounts of /proc and /sys
The initramfs already mounted both of them.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>
---
It's true that this does increase the reliance on an initramfs, but it
is also trivial to revert if Spectrum OS ever stops using one.  Linux
supports creating dm-verity devices via dm-mod.create= and (if needed)
dm-mod.waitfor=, so eliminating an initramfs might be possible but is a
separate project.