tools/sd-notify-adapter: Support exiting if the parent exits
I thought this was impossible to do without races, but it turns out to
be possible.  epoll() will set POLLERR when s6-supervise closes its side
of the pipe.  If FIONREAD returns zero, it means that s6-supervise read
all of the data from the pipe.  This creates a happens-before
relationship between any prctl(PR_SET_PDEATHSIG) and any subsequent
s6-supervise crash.

The signal is specified via the --die-with-parent command-line option.
The default is to not exit if the parent exits.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>
---
The rest of the patches in this series do not use this feature.
However, it cannot be independently implemented, because the parent
death signal is cleared by fork() and so will not be set in the child
process.