scripts/make-erofs.sh: run chmod less
The common case here is that we don't have to change any permissions,
so rewrite it so that (assuming [ is a shell builtin) we don't need to
run any additional programs unless we actually need to make a
permissions change.

I also find this a bit more readable than the previous awk version,
and we no longer need to ignore errors from chmod.

Link: https://spectrum-os.org/lists/archives/spectrum-devel/87plbj4w9q.fsf@alyssa.is
Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-ID: <20250930092249.571803-2-hi@alyssa.is>
Reviewed-by: Demi Marie Obenour <demiobenour@gmail.com>