Generate file lists from a script
Right now, the makefiles in host/rootfs, vm/sys/net, and img/app have
manually-maintained lists of files and symlinks.  These duplicate the
information in the git repository and can easily get out of sync or
cause unnecessary merge conflicts.  Fix all of these issues by having
the git repository be the source of truth, and using a script to
generate the file lists.  Developers can regenerate the lists before
every commit, or even add a git hook to do that.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>
Message-ID: <20251002-genfiles-v9-1-8cf059baf07a@gmail.com>
Reviewed-by: Alyssa Ross <hi@alyssa.is>
Tested-by: Alyssa Ross <hi@alyssa.is>
Signed-off-by: Alyssa Ross <hi@alyssa.is>