tools/xdp-forwarder: Do not include libc headers in eBPF programs
The build happened to work on arm64 because the glibc arm64 headers
don't support multilib.  On x86_64, glibc headers assume that BPF is a
32-bit platform (because __x86_64__ isn't defined) and fail to find the
32-bit headers.  This is not a glibc bug.  Rather, BPF programs should
not be including glibc headers.

Most Linux headers are not trivial to include in BPF programs.  The
version of the headers meant for userspace use do include glibc headers,
and that isn't supported in BPF.  The version meant for building
kernel modules does not, but using it requires much more complicated
build system.

Solve this problem by only including headers intended for use in BPF
programs.  These headers include declarations explicitly intended for
use in BPF programs, so if they do pull in libc headers that is a bug.
This also allows using an unwrapped clang.  Nix's wrapped clang is not
suitable when a target is explicitly specified, so this is another a
bug fix.

This prevents vendoring the example headers, so replace them with a
header that only includes common functionality common to both eBPF
programs.  A single struct including both Ethernet II and 802.1Q headers
is used, massively simplifying the code.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>